pyproject.toml
[tool.poetry]
name = "ankikindle"
version = "0.1.0"
description = "A Flask application to manage vocabularies and interact with Anki"
authors = ["Ian Neville <lanur622@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^2.3.2"
SQLAlchemy = "^1.3.19"
Flask-SQLAlchemy = "^2.4.4"
requests = "^2.28.2"
testcontainers = "^3.7.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
watchdog = "^4.0.0"
selenium = "^4.9.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
